requirements.txt
# requirements.txt
faster-whisper
feedparser
requests
hf_xet
openai-whisper